+++ demo/src/App.jsx
@@ -13,6 +13,7 @@ import Scroll from './sandboxes/leva-scroll/src/App'
 import UI from './sandboxes/leva-ui/src/App'
 import Theme from './sandboxes/leva-theme/src/App'
 import CustomPlugin from './sandboxes/leva-custom-plugin/src/App'
+import LevaTransient from './sandboxes/leva-transient/src/App'
 
 const { styled } = createCss({
   theme: {
@@ -31,13 +32,14 @@ const Page = styled('div', {
 
 const links = {
   'leva-minimal': Minimal,
-  'leva-plugin-spring': PluginSpring,
-  'leva-plugin-plot': PluginPlot,
   'leva-busy': Busy,
   'leva-advanced-panels': AdvancedPanels,
   'leva-scroll': Scroll,
   'leva-ui': UI,
   'leva-theme': Theme,
+  'leva-transient': LevaTransient,
+  'leva-plugin-spring': PluginSpring,
+  'leva-plugin-plot': PluginPlot,
   'leva-custom-plugin': CustomPlugin,
 }