+++ packages/leva/src/types/internal.ts
@@ -3,6 +3,8 @@ import type { SpecialInput, RenderFn, FolderSettings, Plugin } from './public'
 
 export type State = { data: Data }
 
+export type MappedPaths = Record<string, { path: string; onChange: (value: any) => void }>
+
 export type StoreType = {
   useStore: UseStore<State>
   orderPaths: (paths: string[]) => string[]
@@ -19,13 +21,13 @@ export type StoreType = {
   // TODO possibly better type this
   set: (values: Record<string, any>) => void
   get: (path: string) => any
-  getDataFromSchema: (schema: any) => [Data, Record<string, string>]
+  getDataFromSchema: (schema: any) => [Data, MappedPaths]
 }
 
 type Decorators = {
   __refCount: number
   key: string
-  label: string
+  label: string | JSX.Element
   hint?: string
   render?: RenderFn
 }
@@ -35,6 +37,7 @@ export type DataInput = {
   value: unknown
   optional: boolean
   disabled: boolean
+  onChange?: (value: unknown) => void
   settings?: object
 } & Decorators