+++ packages/leva/CHANGELOG.md
@@ -1,5 +1,55 @@
 # leva
 
+## 0.7.3
+
+### Patch Changes
+
+- f9f7f1e: - Have `Select` accept functions as value or options. Fixes #165.
+  - Temporarily type the `set` function values with `any` when `useControls` is used with the function API. In the future, we should infer th value type from the sanitize function.
+
+## 0.7.2
+
+### Patch Changes
+
+- d2eaf58: fix: properly resolves value-keyed objects for the `Select` input.
+
+## 0.7.1
+
+### Patch Changes
+
+- 98984e1: types: fix `set` types in transient mode.
+
+## 0.7.0
+
+### Minor Changes
+
+- 0b7e968: Add the `invertY` setting for the Vector2D joystick for inverting the y coordinate.
+
+  **BREAKING**: The default behavior has been changed. If you want the same behavior as in previous versions you will have to set the `joystick` option to `'invertY'`.
+
+  ```tsx
+  const values = useControl({
+    vector2d: {
+      value: [0, 0],
+      joystick: 'invertY',
+    },
+  })
+  ```
+
+### Patch Changes
+
+- f323cfc: Feat: `onChange` callback for transient updates
+
+  ```js
+  useControls({ color: { value: 'red', onChange: v => console.log(v) } })
+  ```
+
+## 0.6.3
+
+### Patch Changes
+
+- fa909f0: Allow useControls to update settings and input options when dependency array changes. (See #124)
+
 ## 0.6.2
 
 ### Patch Changes